refactor(brief): extract annotation state and overlays from BriefTab
Edit and Delete are quiet links under the speech bubble rather than bordered
buttons, so they stop competing with Reply; the editor's Save/Cancel stay as
buttons because they commit a change.

BriefTab had grown past the cyclomatic limit (26 > 20) as commenting landed,
failing CI. The annotation concern now lives in useBriefAnnotations (selection,
marks, orphan detection, open thread, visibility) and its overlays render from
a BriefAnnotationLayer component, leaving BriefTab as layout. That is also
where the next selection actions will go.
